NONE: Refactored services
