--- renderer/src/components/TimetrackerWebsiteConncetion/index.ts
@@ -1,1 +1,0 @@
-export { default as TimetrackerWebsiteConnection } from "./TimetrackerWebsiteConncetion";
